Update Main.Java
Fixed a lot of things, for example the ANSI parsing. If anyone wants to take a look at it, look at f1(String msg) as it was what is needed to get rid of the ANSI Escape Color Code. Anyways I also added an Icon to my program, to make it a bit more professional, which reminds me, I am now using JFrame instead of JDialog (idk why I was using it), so now it should have a spot on the taskbar.